reworked extension #9
- transferred the extension via dmnd database into refineGEMs as a GeneGapFiller variant
- updated extension to use the functions mentioned above
- reworked configs, cmd_access etc accordingly
- TODO: testing, checking the connections + cleanup still the same
