+++ src/specimen/cmd_access.py
@@ -152,34 +152,32 @@ def refinement():
     """Step 3 of the workflow: Refinement of the model
     """
 
+# @TEST reworked after fixing extension
 @refinement.command()
 @click.option('--draft', type=str, required=True, help='Path to the draft model.')
-@click.option('--gene-list', '-g', type=str, required=True, help='Path to a csv file containing information on all the genes found in the annotated genome.')
+@click.option('--gff', type=str, required=True, help='Path to the gff file of the draft model.')
 @click.option('--fasta', '-f', type=str, required=True, help="Path to the (protein) FASTA file containing the CDS sequences")
 @click.option('--db', '--database', type=str, required=True, help="string, path to the database used for running DIAMOND.")
-# MetaNetX links
-@click.option('--mnx-chem-prop', type=str, required=True, help='Path to the MetaNetX chem_prop namespace file.')
-@click.option('--mnx-chem-xref', type=str, required=True, help='Path to the MetaNetX chem_xref namespace file.')
-@click.option('--mnx-reac-prop', type=str, required=True, help='Path to the MetaNetX reac_prop namespace file.')
-@click.option('--mnx-reac-xref', type=str, required=True, help='Path to the MetaNetX reac_xref namespace file.')
 # NCBI mapping options
 @click.option('--ncbi-map', type=str, required=False, help='Path to the ncbi information mapping file. Optional, but recommended.')
-@click.option('--ncbi-dat', type=str, required=False, help='Path to the ncbi database information file. Optional, but recommended.')
 # DIAMOND and other params
 @click.option('--dir','-d', type=str, default='./refinement/', help='Path to the directory for the output (directories)')
-@click.option('--id', '-i', type=str, default='locus_tag', help='Name of the column of the csv file that contains the entries that were used as gene identifiers in the draft model.')
+@click.option('--mail', '-m', type=str, default=None, help="User's mail address for accessing information from NCBI.")
 @click.option('--sensitivity', '-s', type=click.Choice(['sensitive','more-sensitive','very-sensitive','ultra-sensitive']), default='sensitive', help='Sensitivity mode for DIAMOND blastp run. Default is sensitive.')
 @click.option('--coverage', '-c', type=float, default=80.0, help="Threshold value for the query coverage for DIAMOND. Default is 80.0.")
 @click.option('--pid', type=float, default=95.0, help='PID (percentage identity value) to filter the blast hist by. Default is 90.0, only hits equal or above the given value are kept.')
 @click.option('--threads', '-t', type=int, default=2, help='Number of threads to be used.')
+@click.option('--threshold_add_reacs', '--thres', type=int, default=5, show_default=True, help='Max. number of EC number to reaction mapping cinsidered to be specific enough to add them to the model.')
+@click.option('--prefix', type=str, default='HQTBext', show_default=True, help='Prefix for entity IDs, that have to be build from scratch (fantasy IDs)')
+@click.option('--namespace', '-n', type=click.Choice(['BiGG']), show_default=True, help='Namespace to use for the model IDs. Currently constricted to BiGG by refineGEMs')
 @click.option('--include-dna', is_flag=True, default=False, help='Include reactions with DNA in their name when added (developer information: True == excluded).')
 @click.option('--include-rna', is_flag=True, default=False, help='Include reactions with RNA in their name when added (developer information: True == excluded).')
 @click.option('--memote', is_flag=True, default=False, help='Use memote on the extended model.')
-def extension(draft, gene_list, fasta, db, dir,
-    mnx_chem_prop, mnx_chem_xref, mnx_reac_prop, mnx_reac_xref,
-    ncbi_map, ncbi_dat,
-    id, sensitivity,
+def extension(draft, gff, fasta, db, dir,
+    ncbi_map, mail,
+    sensitivity,
     coverage, pid, threads,
+    threshold_add_reacs, prefix, namespace, 
     include_dna, include_rna,
     memote):
     """Refinement step 1: Extend the model.
@@ -188,11 +186,11 @@ def extension(draft, gene_list, fasta, db, dir,
     draft, gene_list, fasta, db, dir,
     mnx_chem_prop, mnx_chem_xref, mnx_reac_prop, mnx_reac_xref
     """
-    specimen.hqtb.core.refinement.extension.run(draft, gene_list, fasta, db, dir,
-        mnx_chem_prop, mnx_chem_xref, mnx_reac_prop, mnx_reac_xref,
-        ncbi_map, ncbi_dat,
-        id, sensitivity,
+    specimen.hqtb.core.refinement.extension.run(draft, gff, fasta, db, dir,
+        ncbi_map, mail,
+        sensitivity,
         coverage, pid, threads,
+        threshold_add_reacs, prefix, namespace,
         not include_dna, not include_rna,
         memote)
