+++ src/specimen/data/config/hqtb_config_default.yaml
@@ -16,6 +16,7 @@
 # information about the genome to be used to generate the new model
 subject:
     annotated_genome: __USER__
+    gff: __USER___
     full_sequence: __USER__
 
 # information about the template model/genome
@@ -41,14 +42,10 @@ data:
     data_direc: null
     # required
     diamond: __USER__
-    # needed but potentially downloaded
-    mnx_chem_prop: MetaNetX/chem_prop.tsv
-    mnx_chem_xref: MetaNetX/chem_xref.tsv
-    mnx_reac_prop: MetaNetX/reac_prop.tsv
-    mnx_reac_xref: MetaNetX/reac_xref.tsv
+
     # optional, but good and manual
     ncbi_map: null
-    ncbi_dat: null
+
     # optional for directionality control
     biocyc: null
     # optional:
@@ -62,6 +59,12 @@ data:
 
 # paramters for the single steps of the pipeline
 parameters:
+    general:
+        idprefix: 'HQTB'    # prefix to use for fantasy IDs, if IDs for 
+                            # the namespace do not exist.
+        namespace: 'BiGG'
+        email: USER         # email address. Required for queries against NCBI
+
     bidirectional_blast:
         # default should suffice except special cases
         template_name: null
@@ -77,17 +80,22 @@ parameters:
         medium: default
 
     refinement_extension:
-        # default (usually) fine
-        id: null
-        # default fine
-        sensitivity: more-sensitive
-        # default alright but good to edit for trying different options
-        coverage: 95.0
-        pid: 90.0
-        # default almost needed, except for special cases
-        exclude_dna: True
-        exclude_rna: True
-        email: USER
+        # params for DIAMOND
+        sensitivity: more-sensitive # change the sensitivity of DIAMOND
+        coverage: 95.0              # Set the min. coverage 
+        pid: 90.0                   # set the minimal percentage identify value for valid hits
+        # params for adding entities during 
+        formula-check: 'existence'  # When checking, if a metabolite can be added to the model
+                              # also check the formula. For more information about
+                              # available options, please refer to the docs of 
+                              # the function isreaction_comlete().
+        exclude-dna: True           # Exclude reactions containing 'DNA' in their name
+                                    # from being added to the model.
+        exclude-rna: True           # Exclude reactions containing 'RNA' in their name
+                                    # from being added to the model.
+        threshold_add_reacs: 5      # Threshold values. If the mapping of an EC against reactions
+                                    # exceeds this value, the reactions will not be added, due to 
+                                    # the high probability of false positives. 
 
     refinement_cleanup:
         # default as standard
