+++ src/specimen/hqtb/core/refinement/smoothing.py
@@ -3,6 +3,8 @@
 
 __author__ = 'Carolin Brune'
 
+# @TODO logging 
+
 ################################################################################
 # requirements
 ################################################################################
